Function enable_config::main

pub(crate) fn main() -> Result<(), Error>

Enables AWS Config for a resource type, in the Region.

Arguments

-b BUCKET - The name of the Amazon bucket to which AWS Config delivers configuration snapshots and configuration history files.

-i IAM-ARN - The ARN of the IAM role that used to describe the AWS resources associated with the account.

-k KMS-ARN - The ARN of the KMS key that used to encrypt the data in the bucket.

-p PREFIX - The prefix for the bucket.

-s SNS-ARN - The ARN of the Amazon SNS topic to which AWS Config sends notifications about configuration changes.

[-t TYPE] - The type of resource for AWS Config to support. If not supplied, defaults to AWS::DynamoDB::Table (DynamoDB tables).

[-n NAME] - The name of the configuration. If not supplied, defaults to default.

[-r REGION] - The Region in which the client is created. If not supplied, uses the value of the AWS_REGION environment variable. If the environment variable is not set, defaults to us-west-2.

[-v] - Whether to display information. Need: s3 key prefix AND kms key